address comments from review
