+++ crypto/algorithm_test.go
@@ -6,11 +6,60 @@ import (
 
 	"github.com/stretchr/testify/require"
 
+	ethcrypto "github.com/ethereum/go-ethereum/crypto"
+
 	"github.com/cosmos/cosmos-sdk/crypto/keyring"
 	"github.com/cosmos/cosmos-sdk/crypto/keys/hd"
 	"github.com/cosmos/cosmos-sdk/tests"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
+
+func TestEthermintKeygenFunc(t *testing.T) {
+	privkey, err := GenerateKey()
+	require.NoError(t, err)
+
+	testCases := []struct {
+		name    string
+		privKey []byte
+		algo    keyring.SigningAlgo
+		expPass bool
+	}{
+		{
+			"valid ECDSA privKey",
+			ethcrypto.FromECDSA(privkey.ToECDSA()),
+			EthSecp256k1,
+			true,
+		},
+		{
+			"nil bytes, valid algo",
+			nil,
+			EthSecp256k1,
+			true,
+		},
+		{
+			"empty bytes, valid algo",
+			[]byte{},
+			EthSecp256k1,
+			true,
+		},
+		{
+			"invalid algo",
+			nil,
+			keyring.MultiAlgo,
+			false,
+		},
+	}
+
+	for _, tc := range testCases {
+		privkey, err := EthermintKeygenFunc(tc.privKey, tc.algo)
+		if tc.expPass {
+			require.NoError(t, err, tc.name)
+		} else {
+			require.Error(t, err, tc.name)
+			require.Nil(t, privkey, tc.name)
+		}
+	}
+}
 
 func TestKeyring(t *testing.T) {
 	dir, cleanup := tests.NewTestCaseDir(t)
