--- crypto/keys_test.go
@@ -1,58 +1,0 @@
-package crypto
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/require"
-
-	ethcrypto "github.com/ethereum/go-ethereum/crypto"
-
-	"github.com/cosmos/cosmos-sdk/crypto/keyring"
-)
-
-func TestEthermintKeygenFunc(t *testing.T) {
-	privkey, err := GenerateKey()
-	require.NoError(t, err)
-
-	testCases := []struct {
-		name    string
-		privKey []byte
-		algo    keyring.SigningAlgo
-		expPass bool
-	}{
-		{
-			"valid ECDSA privKey",
-			ethcrypto.FromECDSA(privkey.ToECDSA()),
-			EthSecp256k1,
-			true,
-		},
-		{
-			"nil bytes, valid algo",
-			nil,
-			EthSecp256k1,
-			true,
-		},
-		{
-			"empty bytes, valid algo",
-			[]byte{},
-			EthSecp256k1,
-			true,
-		},
-		{
-			"invalid algo",
-			nil,
-			keyring.MultiAlgo,
-			false,
-		},
-	}
-
-	for _, tc := range testCases {
-		privkey, err := EthermintKeygenFunc(tc.privKey, tc.algo)
-		if tc.expPass {
-			require.NoError(t, err, tc.name)
-		} else {
-			require.Error(t, err, tc.name)
-			require.Nil(t, privkey, tc.name)
-		}
-	}
-}
